+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `Request::with_param` for more ergonomic query parameter
+  usage. Thanks for the PR, @sjvignesh!
+  ([#54](https://github.com/neonmoe/minreq/pull/54))
+
 ### Fixed
 - Chunk length handling for some servers with slightly off-spec chunk
   lengths. Thanks for the report, @Shnatsel!