+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `Request::with_max_status_line_length` for avoiding DoS when the
   server sends large headers or status lines. Thanks for the report,
   @Shnatsel! ([#55](https://github.com/neonmoe/minreq/issues/55))
+- Support for the `rustls-native-certs` crate via a new
+  `https-rustls-probe` feature. Thanks for the PR, @joeried!
+  ([#59](https://github.com/neonmoe/minreq/pull/59))
 
 ### Fixed
 - Chunk length handling for some servers with slightly off-spec chunk